requirements.txt
pytorch
py-rouge
tqdm
nltk
num2words
